Automatically generate tests for checkpoint versionning (#669)
